Update buildCleanup.groovy
Changed the cron to run once a day instead of every minute. 
Moved the 'def config' into the 'jobs' method as this is only needed incase of crone jobs and not for exceution - this was throwing an exception in case that the properties file was not exist.
Uncommented the 'builds.deleteBuild it' as it prevented the actual deletion.