Added empty test files
All user plugins in this repository should have a test file, which may
be empty. This is useful because it allows the devenv to link in a test
that may not exist, so that it can be written (the alternative is to
either create the test before linking, relink after creating the test,
or link the test manually).

Some new plugins have been added without empty tests. This adds them.