Fixed bug in ldapGroupsConfig
An LDAP group with a null LDAP setting link causes a NPE in some
situations. Noticed that groups with no setting links are internally
stored with a string containing a single space in that field when done
from the web UI, so changed the config plugin to store them that way.