Fixed statusPropertyWorkflow to load in 4.x
The statusPropertyWorkflow plugin had been using an internal API that
hasn't existed since Artifactory 3.9, but now it's been updated to use
the new 4.x API instead. The rest of this plugin still needs to be
tested, as I don't know enough about Blackduck to get it running and get
this tested quickly.