Fixed bug in artifactCleanup plugin
The artifactCleanup plugin could potentially 404 if it tries to delete
an artifact that does not exist. This could happen if the plugin finds
an artifact to clean, but the artifact is removed before the plugin
could get around to cleaning it. This is increasingly likely since the
pause feature was added. Now, if the plugin reaches an artifact that
doesn't exist, it just moves on to the next artifact.