Minor fix for yumValidRemote plugin
Before, yumValidRemote would replace all repomd.xml files, even if the
xml was unmodified from its original state. This could potentially cause
checksum conflicts (if, say, the whitespace is different), which can
cause smart remotes to complain. Now, if the plugin realizes that none
of the fields are going to change, it just exits early.