Attempt to fix requestRouting issue
The requestRouting plugin was using a Predicate 'heartbeat' defined in
Artifactory, and calling it as if it was just a regular function. This
should work fine, but it doesn't appear to in all cases. Now, the
Predicate is not called anymore, and the calculation of the heartbeat is
instead implemented directly.

Change-Id: I914ff1922c2f8c8b2d6c4ee429d025e133f34767