Fixed small bugs in securityReplication
When another instance distributes the json config file via the
securityReplication endpoint to an HA instance, it may go to a secondary
node, which would cause it to not be copied to the primary node. Now,
the securityReplication endpoint redirects this request to the primary
node in this case.

When the cron job triggers, the plugin would first ping all the
instances until it found a running instance (the master), and then if it
was the master it would ping all the instances again. Now, these two
steps are combined into one, so each instance is only pinged a maximum
of once by each other instance.

Change-Id: I3b35077e37f32586fe9830bca8c1677ff2f53ade