Moved buildArtifactsAGVList around
In build/mavenBuildDeps, there used to be a user plugin, a maven plugin
which used it, and an example of using the two together. There has been
a fair amount of confusion about this, so I've updated the file
structure. Now, the user plugin is stored in
build/buildArtifactsAGVList, which contains a util folder containing the
maven plugin and example.