Fixed yumValidRemote for Artifactory 5.6
Changed the cache file rename step to use a different API that bypasses
validations, as those assume we're in a local repo, and not a remote
cache. In the new version of Artifactory, this causes an NPE.

Added more logging. When a spawned thread reaches an exception, by
default it isn't logged, so we have to catch the exception and log it
manually.