Fixed securityReplication bug
Fixed a bug in securityReplication relating to data loss when the filter
levels are mismatched between the nodes in the mesh. Now, the master
node sends its own filter levels to the other nodes, and those are used
instead. This way, all nodes use the master's filter level, so no
mismatch can take place.