More reliable distSecRep
The distSecRep and securityReplication endpoints used to distribute to
the primary node of a cluster only, and depend on Artifactory's internal
plugin file management system to move files to the other nodes. However,
Artifactory does not appear to do this very consistently. Now,
distSecRep distributes to all nodes in all clusters directly, bypassing
the issue.