Fixed minor bug in securityReplication
Permissions that refer to remote repos actually refer to the local
cache. This can cause a permission to not apply if it's replicated from
an instance where the repo it refers to is remote to an instance where
the repo is local. This is now fixed, as the '-cache' suffix is removed
when permissions are read from the database.