Added 'acceptableFailures' variable
When debugging the securityReplication plugin, it can be helpful to skip
a number of write failures (say, if a user is corrupted, for the sake of
debugging the plugin, it might be useful if the corrupted user did not
abort the entire replication cycle). This modification allows a maximum
number of "acceptable failures" to occur before the cycle aborts.