Added configurable time units
Until now as time unit only "month" have been supported. This has prevented using the plugin for cleaning up the artifacts that have been not downloaded for e.g. the last 14 days. Therefore the new parameters "timeUnit" and "timeInterval" are introduced, while the support for the old "months" is still preserved. This allows a far more flexible configuration when artifacts can be deleted. Moreover the cleanup policies for different repositories defined in the "artifactCleanup.properties" are still supported, but there is a new json-based "artifactCleanup.json" introduced that is allowing a way more transparent and extensible policies definition. The introduction of the new file also allows to stay backward compatible and avoids building in not maintainable code. The README.md has been updated accordingly to these changes.