Updated synchronizeLdapGroups plugin
Now the plugin ensures that LDAP actually has groups to add before
adding them to the groups list. It does this by removing all groups
already in the groups list from LDAP's list, and then checking if the
result is empty. This way, the update only happens when absolutely
necessary.